+++ collector/cluster_node.go
@@ -6,7 +6,7 @@ import "github.com/prometheus/client_golang/prometheus"
 var (
 	soraClusterMetrics = SoraClusterMetrics{
 		clusterNode: newDescWithLabel("cluster_node", "The sora server known cluster node.", []string{"node_name", "mode"}),
-		raftRole:        newDescWithLabel("cluster_raft_role", "The current Raft role. The role name is indicated by the label 'role'. The value of this metric is always set to 1.", []string{"role"}),
+		raftState:        newDescWithLabel("cluster_raft_state", "The current Raft state. The state name is indicated by the label 'state'. The value of this metric is always set to 1.", []string{"state"}),
 		raftTerm:        newDesc("cluster_raft_term", "The current Raft term."),
 		raftCommitIndex: newDesc("cluster_raft_commit_index", "The latest committed Raft log index."),
 	}
@@ -14,14 +14,14 @@ var (
 
 type SoraClusterMetrics struct {
 	clusterNode     *prometheus.Desc
-	raftRole        *prometheus.Desc
+	raftState        *prometheus.Desc
 	raftTerm        *prometheus.Desc
 	raftCommitIndex *prometheus.Desc
 }
 
 func (m *SoraClusterMetrics) Describe(ch chan<- *prometheus.Desc) {
 	ch <- m.clusterNode
-	ch <- m.raftRole
+	ch <- m.raftState
 	ch <- m.raftTerm
 	ch <- m.raftCommitIndex
 }
@@ -34,7 +34,7 @@ func (m *SoraClusterMetrics) Collect(ch chan<- prometheus.Metric, nodeList []sor
 			ch <- newGauge(m.clusterNode, 1, *node.NodeName, *node.Mode)
 		}
 	}
-	ch <- newGauge(m.raftRole, 1.0, report.RaftRole)
+	ch <- newGauge(m.raftState, 1.0, report.RaftState)
 	ch <- newCounter(m.raftTerm, float64(report.RaftTerm))
 	ch <- newCounter(m.raftCommitIndex, float64(report.RaftCommitIndex))
 }